requirements.txt
asgiref==3.7.2
Django==4.2.2
django-cors-headers==4.1.0
django-jazzmin==2.6.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
openpyxl==3.1.2
Pillow==9.5.0
python-monkey-business==1.0.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
